[cdac] Add IManagedTypeLayout contract for metadata-driven managed type layouts
Adds a new cDAC contract, IManagedTypeLayout, that resolves managed type
layouts (instance field offsets and static field addresses) from ECMA
metadata via IRuntimeTypeSystem instead of from hardcoded native data
descriptor entries.

Following the existing ConditionalWeakTable precedent, all managed type
layouts consumed by the cDAC are now read from runtime metadata. Callers
look up a type once (by namespace + name) to get a ManagedTypeInfo with
field offsets pre-shifted by sizeof(Object) so the standard
`address + field.Offset` idiom works for reference-type reads.

Contract: ManagedTypeLayout, version c1.

Migrated consumers:
* ConditionalWeakTable_1
* SyncBlock_1
* ComWrappers_1
* Object_1 / Data.String, Data.Exception,
  Data.NativeObjectWrapperObject, Data.ManagedObjectWrapperHolderObject

Other notes:
* Data.String.StringLength is now `int` to match the managed field type.
* SyncBlock_1 reads `Lock._owningThreadId` as `int` (matches the managed
  field) and casts to `uint` for the existing API surface.
* Removes the corresponding managed-type entries from datadescriptor.inc
  and adds the ManagedTypeLayout contract registration.
* Adds docs/design/datacontracts/ManagedTypeLayout.md and updates the
  affected contract docs.

Validated locally:
* cDAC unit tests: 1691/1691 passing
* cDAC dump tests: 215 passed / 0 failed (135 environment skips)

+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Contracts/Contracts/Object_1.cs
@@ -50,7 +50,7 @@ internal readonly struct Object_1 : IObject
         if (str.StringLength == 0)
             return string.Empty;
 
-        Span<byte> span = stackalloc byte[(int)str.StringLength * sizeof(char)];
+        Span<byte> span = stackalloc byte[str.StringLength * sizeof(char)];
         _target.ReadBuffer(str.FirstChar, span);
         return new string(MemoryMarshal.Cast<byte, char>(span));
     }
